✨ userGroupRelation.tsを作成
