+++ src/mocks/handlers.ts
@@ -1,37 +1,10 @@
 import { graphql } from 'msw'
-import { UserGroupRelation } from '~/types/generated/graphql'
-import {
-  favoriteFactory,
-  userFactory,
-  userGroupRelationFactory,
-} from './factories'
+import { getUser, getUserGroupRelations } from './api'
+import { favoriteFactory } from './factories'
 
 export const handlers = [
-  graphql.query('GetUserGroupRelations', (req, res, ctx) => {
-    const { userId, groupId } = req.variables
-    const query: Partial<UserGroupRelation> = {}
-    if (userId) query.userId = userId
-    if (groupId) query.groupId = groupId
-    return res(
-      ctx.data({
-        GetUserGroupRelations: [
-          {
-            ...userGroupRelationFactory(query),
-          },
-        ],
-      })
-    )
-  }),
-  graphql.query('GetUser', (req, res, ctx) => {
-    const { getUserId } = req.variables
-    return res(
-      ctx.data({
-        getUser: {
-          ...userFactory({ id: getUserId, image: 'images/Desii_icon.png' }),
-        },
-      })
-    )
-  }),
+  graphql.query('GetUserGroupRelations', getUserGroupRelations),
+  graphql.query('GetUser', getUser),
   graphql.query('GetFavorites', (req, res, ctx) => {
     const { postId } = req.variables
     return res(
